fix/rename mock data file to resolve import issue
